PZ_4/backend/requirements.txt
# API Framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.2
pydantic-settings==2.1.0
email-validator==2.0.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1

# Data Processing
pandas==2.1.3
numpy==1.26.2
scipy==1.10.1

# Machine Learning
scikit-learn==1.3.2
tensorflow==2.12.0
keras==2.12.0
xgboost==2.0.2
lightgbm==4.1.0

# Visualization
plotly==5.18.0
matplotlib==3.8.2
seaborn==0.13.0

# Testing
pytest==7.3.1
pytest==6.2.5
pytest-cov==4.1.0
httpx==0.19.0

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# ML Experiment Tracking
mlflow==2.8.1

# Utilities
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.9.1
pandas-ta==0.3.14b0  # Technical analysis for financial data
yfinance==0.2.33     # Yahoo Finance API 